Converted most of WhileyWeb
Am in the process of converting WhileyWeb from a Python application to a
Java application based on Apache HttpComponents.  The benefits of doing
this are mostly around the ability to run the Whiley Compiler in memory.

At the moment, this code is pretty rubbish.  Mostly, this is a learning
experience for me to: firstly, learn Apache HttpComponents; secondly, to
figure out what we need to make this thing work.