Interfacing with WhileyCompiler
Have added a very rudimentary interface to the WhileyCompiler.  This
compiles a single file, generates error messages, etc.  At the moment,
it is a complete hack which is just for me to see the whole pipeline.

At the moment, verification isn't working.  The problem is related to
the interface between VirtualRool and WhileyFileParser (roughly
speaking).