Add build.xml
Added an initial build.xml file for building WhileyLabs.  This pulls
down the necessary apache dependencies from Maven Central, and compiles
the relevant source files.

The build.xml also provides a rudimentary mechanism for running the
server.  This is helpful because it includes the necessary dependencies
on the classpath, rather than having to specify them explicitly.
However, it's not actually working right now.