Updated to latest version of WhileyCompiler
This updates the Web Application to define its dependencies via a
pom.xml (this helps with loading the project into Eclipse).  It also
pulls all dependencies which form the WhileyCompiler, whereas previously
they were stored in a local jar.

The compiler itself is updated to support the latest API, though it's
not hugely pretty right now.  But, it does work and will correctly build
and verify programs, and report error messages as necessary.  However,
it is not currently including the standard library as part of the
bootpath.  This is obviously a problem in the long run!