Basic support for ACE
This puts in place basic support for the ACE editor using the bindings
in ACE.wy.  Seems to work pretty well and quite easily.y

At this stage, can now see error annotations in the left gutter.
However, to get the underlining of errors requires being able to use
EditSession::addMarker.

There is essentially a problem with the EditSession.getMarkers() method as it
doesn't do what is says on the tin.  Specifically, it doesn't return an array!
Instead, it returns an object of some kind for reasons unknown.  I think the
only way to remove markers is to remember their IDs.

Messages starting to work

This updates the POM.xml file to the latest version of Whiley.  However, it no
longer compiles!

Now running again.

There are quite a lot of problems with the organisation of Build.Environments,
Command.Environments, Configurations and all that.  It's just a total mess!