Working on Compilation Config
Basic plan is to finish recreating what I've already got.  Then, to convert into
a functional style which will be the easiest to maintain in the long run.

This adds support for verification and counter examples.  At this stage,
I think that's enough in terms of converting from JavaScript to Whiley.