Update Dockerfile
This updates Dockerfile so that it works again!!  This was a little bit more
involved than expected because got sidetracked with the fact that Maven central
stopped accepting http, and only accepts https.

Have also simplified the Dockerfile so that it's based of openjdk:8-alpine which
is a bit nicer.