Initial conversion to Web.wy
This puts through an initial convertion to the FRP framework Web.wy.  At
this stage, it does read information from the ACE editor and post it to
the server.  It also picks up the response, but only produces an alert.

The next step is to process the server response properly.  This requires
turning the response back into a Whiley object, and also implementing
the message box.