Abandon support for code execution; fix pkgs
This abandons any effort to support code execution at this stage.  Firstly, it's
unsafe.  Secondly, i couldn't figure out how to make it work even though it is
technically possible.

This also fixes package support, so that it will now compile against the
standard library provided that is available in the local repository when you run
the server.